feat: reorganizar la conexión a ElasticSearch y agregar script para indexar libros
